requirements.txt
torch==2.8.0
transformers==4.56.0
datasets==4.0.0
tokenizers==0.22.0
accelerate==1.10.1

scikit-learn==1.7.1
numpy==2.3.2
pandas==2.3.2

# Utilities
tqdm==4.67.1
pyyaml==6.0.2
